introduce-weight-updater-controller(extract-move): Hand disk-based weight reload over to WeightUpdaterController
Pure physical move per MECH_COMMIT_SPLIT. Cut the @staticmethod methods
``update_weights_from_disk``, ``_wait_for_model_update_from_disk`` and
``handle_update_weights_from_disk_req_output`` from TokenizerManager
(leaving ``_update_model_path_info`` on TM) and
``_update_weight_version_if_provided`` from TokenizerControlMixin; paste
into WeightUpdaterController (drop @staticmethod, replace
``self: "WeightUpdaterController"`` → plain ``self``). Flip the
``init_request_dispatcher`` entry from the prep-stage lambda forwarder
to a direct bound-method reference. Caller prefix replacement:
``TokenizerManager.<method>(self.weight_updater_controller, ...)`` →
``self.weight_updater_controller.<method>(...)`` (TM + mixin sibling
methods + entrypoints); ditto for the
``TokenizerControlMixin._update_weight_version_if_provided`` call sites.

Refactor chain ID: introduce-weight-updater-controller-move
